pcb stencil jig with suction air holes
3d printable jig that is designed to help for keeping
the pcb and stencil on place when spreading the solder paste.
This jig will use the suction air that can be get for example
for the cleaning vacuum to keep the stencil on place.

Project can generate all required pieces and it's size is
easily adjustable for different size of pcb boards.
Following 3d models can be generated by using this project:

- container_box that provides the base for which all other parts will
  be layered. All four corners will in addition require 35m length
springs for lifting the pcb_stencil_liftboard up once air suction/vacuum
cleaner is turned off. I have used 0.8x9mm OD,35mm Free Length springs
ordered from Amazon:
https://www.amazon.com/gp/product/B0B6J19Q14/ref=ppx_yo_dt_b_search_asin_title?ie=UTF8&th=1
- lidboard with air holes that is put on top of the container box
- pcb_stencil_liftboard provides a frame that can lower and raise the
  pcb stencil. It is put on top of the lidboard. Air suction holes on pcb_holder_board are
aligned with the iidboard so that the suction air will hold the stencil on place
- pcb_holder_board will fit inside the pcb_stencil_liftboard and it will
  hold the pcb on place. Air suction holes on pcb_holder_board are
aligned with the iidboard so that the suction air will hold the pcb and
stencil on place

Project is written with openscad so the size of the all 3d models
required can be easily adjusted by changing the parameter values in
code. In addition the openscad will quarantee that the project will work
on multiple platforms. (Linux, Mac, Windows, etc...)

Signed-off-by: Mika Laitio <lamikr@gmail.com>